Add "config set (organization|project) id"
The command is used to set a org/project ID in the current context.

- Any suggestion on the command path is appreciated
- Have "config" and "configs" package to avoid cyclic dependency, but
  this need a better package name.
