+++ internal/configs/configs.go
@@ -5,10 +5,22 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 
-	"github.com/spf13/cobra"
+	"github.com/AlecAivazis/survey/v2"
+	"github.com/llm-operator/cli/internal/ui"
 	"github.com/zchee/go-xdgbasedir"
 	"gopkg.in/yaml.v2"
+)
+
+const (
+	configVersion    = "v1"
+	authClientID     = "llm-operator"
+	authClientSecret = "ZXhhbXBsZS1hcHAtc2VjcmV0"
+	authRedirectURI  = "http://127.0.0.1:5555/callback"
+
+	defaultEndpointURL = "http://localhost:8080/v1"
+	defaultIssuerURL   = "http://kong-proxy.kong/v1/dex"
 )
 
 // Auth is an authentication configuration.
@@ -68,12 +80,28 @@ func (c *C) Validate() error {
 	return nil
 }
 
+// Save saves the config.
+func (c *C) Save() error {
+	path := configFilePath()
+	if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
+		return fmt.Errorf("mkdir all: %s", err)
+	}
+	b, err := yaml.Marshal(c)
+	if err != nil {
+		return fmt.Errorf("marshal: %s", err)
+	}
+	if err := os.WriteFile(path, b, 0644); err != nil {
+		return fmt.Errorf("write file: %s", err)
+	}
+	return nil
+}
+
 // LoadOrCreate loads the config.
 func LoadOrCreate() (*C, error) {
 	// Create a config file if it doesn't exists.
 	path := configFilePath()
 	if _, err := os.Stat(path); os.IsNotExist(err) {
-		if err := create(); err != nil {
+		if err := CreateNewConfig(); err != nil {
 			return nil, fmt.Errorf("create default config: %s", err)
 		}
 	}
@@ -90,18 +118,95 @@ func LoadOrCreate() (*C, error) {
 	return &config, nil
 }
 
-func configFilePath() string {
-	return filepath.Join(xdgbasedir.ConfigHome(), "llmo", "config.yaml")
+// CreateNewConfig creates a new config file.
+func CreateNewConfig() error {
+	p := ui.NewPrompter()
+	endpointURL, err := askWithDefaultValue(
+		p,
+		"Input endpoint URL of LLM service",
+		defaultEndpointURL,
+		func(v string) error {
+			if !isHTTPURL(v) {
+				return errors.New("must start with 'http://' or 'https://'")
+			}
+			if !strings.HasSuffix(v, "/v1") {
+				return errors.New("must end with '/v1'")
+			}
+			return nil
+		},
+	)
+	if err != nil {
+		return err
+	}
+
+	issuerURL, err := askWithDefaultValue(
+		p,
+		"Input OIDC Issuer URL for login",
+		defaultIssuerURL,
+		func(v string) error {
+			if !isHTTPURL(v) {
+				return errors.New("must start with 'http://' or 'https://'")
+			}
+			return nil
+		},
+	)
+	if err != nil {
+		return err
+	}
+
+	c := &C{
+		Version:     configVersion,
+		EndpointURL: endpointURL,
+		Auth: Auth{
+			ClientID:     authClientID,
+			ClientSecret: authClientSecret,
+			RedirectURI:  authRedirectURI,
+			IssuerURL:    issuerURL,
+		},
+	}
+	if err := c.Save(); err != nil {
+		return err
+	}
+
+	return nil
 }
 
-// Cmd returns a new config command.
-func Cmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:                "config",
-		Short:              "Config commands",
-		Args:               cobra.NoArgs,
-		DisableFlagParsing: true,
+func askWithDefaultValue(
+	p ui.Prompter,
+	msg string,
+	defaultVal string,
+	validate func(v string) error,
+) (string, error) {
+	var val string
+	if err := p.Ask(
+		&survey.Input{Message: fmt.Sprintf("%s (default: %q):", msg, defaultVal)},
+		&val,
+		survey.WithValidator(func(ans interface{}) error {
+			v, ok := ans.(string)
+			if !ok {
+				return fmt.Errorf("invalid input")
+			}
+			v = strings.TrimSpace(v)
+			if v == "" {
+				return nil
+			}
+
+			return validate(v)
+		}),
+	); err != nil {
+		return "", err
 	}
-	cmd.AddCommand(createCmd())
-	return cmd
+	if val != "" {
+		return val, nil
+	}
+
+	return defaultVal, nil
+}
+
+func isHTTPURL(s string) bool {
+	return strings.HasPrefix(s, "http://") || strings.HasPrefix(s, "https://")
+}
+
+func configFilePath() string {
+	return filepath.Join(xdgbasedir.ConfigHome(), "llmo", "config.yaml")
 }
